+++ core/src/main/java/tech/ydb/core/grpc/GrpcTransport.java
@@ -6,7 +6,6 @@ import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CancellationException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.function.BiConsumer;
@@ -80,9 +79,6 @@ public abstract class GrpcTransport implements AutoCloseable {
         this.discoveryMode = builder.getDiscoveryMode();
     }
 
-    @Nullable
-    protected abstract Channel getChannel();
-
     public static GrpcTransportBuilder forHost(String host, int port) {
         return new GrpcTransportBuilder(null, null, singletonList(HostAndPort.fromParts(host, port)));
     }
@@ -155,17 +151,18 @@ public abstract class GrpcTransport implements AutoCloseable {
         CompletableFuture<Result<RespT>> promise = new CompletableFuture<>();
 
         if (!shutdown) {
-            final Channel channel = getChannel();
-            ClientCall<ReqT, RespT> call = channel.newCall(method, callOptions);
-            if (logger.isDebugEnabled()) {
-                logger.debug("Sending request to {}, method `{}', request: `{}'", channel.authority(), method, request );
-            }
-            sendOneRequest(call, request, new UnaryStreamToFuture<>(promise));
+            return makeUnaryCall(method, request, callOptions, promise);
         } else {
             promise.complete(CancelResultDueToShutdown());
         }
         return promise;
     }
+
+    protected abstract <ReqT, RespT> CompletableFuture<Result<RespT>> makeUnaryCall(
+            MethodDescriptor<ReqT, RespT> method,
+            ReqT request,
+            CallOptions callOptions,
+            CompletableFuture<Result<RespT>> promise);
 
     public <ReqT, RespT> void unaryCall(
             MethodDescriptor<ReqT, RespT> method,
@@ -185,16 +182,17 @@ public abstract class GrpcTransport implements AutoCloseable {
         }
 
         if (!shutdown) {
-            final Channel channel = getChannel();
-            ClientCall<ReqT, RespT> call = channel.newCall(method, callOptions);
-            if (logger.isDebugEnabled()) {
-                logger.debug("Sending request to {}, method `{}', request: `{}'", channel.authority(), method, request);
-            }
-            sendOneRequest(call, request, new UnaryStreamToConsumer<>(consumer));
+            makeUnaryCall(method, request, callOptions, consumer);
         } else {
             consumer.accept(CancelResultDueToShutdown());
         }
     }
+
+    protected abstract <ReqT, RespT> void makeUnaryCall(
+            MethodDescriptor<ReqT, RespT> method,
+            ReqT request,
+            CallOptions callOptions,
+            Consumer<Result<RespT>> consumer);
 
     public <ReqT, RespT> void unaryCall(
             MethodDescriptor<ReqT, RespT> method,
@@ -214,16 +212,17 @@ public abstract class GrpcTransport implements AutoCloseable {
         }
 
         if (!shutdown) {
-            final Channel channel = getChannel();
-            ClientCall<ReqT, RespT> call = channel.newCall(method, callOptions);
-            if (logger.isDebugEnabled()) {
-                logger.debug("Sending request to {}, method `{}', request: `{}'", channel.authority(), method, request );
-            }
-            sendOneRequest(call, request, new UnaryStreamToBiConsumer<>(consumer));
+            makeUnaryCall(method, request, callOptions, consumer);
         } else {
             consumer.accept(null, Status.CANCELLED);
         }
     }
+
+    protected abstract <ReqT, RespT> void makeUnaryCall(
+            MethodDescriptor<ReqT, RespT> method,
+            ReqT request,
+            CallOptions callOptions,
+            BiConsumer<RespT, Status> consumer);
 
     public <ReqT, RespT> StreamControl serverStreamCall(
             MethodDescriptor<ReqT, RespT> method,
@@ -243,17 +242,18 @@ public abstract class GrpcTransport implements AutoCloseable {
         }
 
         if (!shutdown) {
-            final Channel channel = getChannel();
-            ClientCall<ReqT, RespT> call = channel.newCall(method, callOptions);
-            sendOneRequest(call, request, new ServerStreamToObserver<>(observer, call));
-            return () -> {
-                call.cancel("Cancelled on user request", new CancellationException());
-            };
+            return makeServerStreamCall(method, request, callOptions, observer);
         } else {
             observer.onError(CancelResultDueToShutdown().toStatus());
             return () -> {};
         }
     }
+
+    protected abstract <ReqT, RespT> StreamControl makeServerStreamCall(
+            MethodDescriptor<ReqT, RespT> method,
+            ReqT request,
+            CallOptions callOptions,
+            StreamObserver<RespT> observer);
 
     private <ReqT> OutStreamObserver<ReqT> makeEmptyObserverStub() {
         return new OutStreamObserver<ReqT>() {
@@ -288,16 +288,19 @@ public abstract class GrpcTransport implements AutoCloseable {
         }
 
         if (!shutdown) {
-            final Channel channel = getChannel();
-            ClientCall<ReqT, RespT> call = channel.newCall(method, callOptions);
-            return asyncBidiStreamingCall(call, observer);
+            return makeBidirectionalStreamCall(method, callOptions, observer);
         } else {
             observer.onError(CancelResultDueToShutdown().toStatus());
             return makeEmptyObserverStub();
         }
     }
 
-    private static <ReqT, RespT> void sendOneRequest(
+    protected abstract <ReqT, RespT> OutStreamObserver<ReqT> makeBidirectionalStreamCall(
+            MethodDescriptor<ReqT, RespT> method,
+            CallOptions callOptions,
+            StreamObserver<RespT> observer);
+
+    protected static <ReqT, RespT> void sendOneRequest(
             ClientCall<ReqT, RespT> call,
             ReqT request,
             ClientCall.Listener<RespT> listener) {
@@ -317,7 +320,7 @@ public abstract class GrpcTransport implements AutoCloseable {
         }
     }
 
-    private static <ReqT, RespT> OutStreamObserver<ReqT> asyncBidiStreamingCall(
+    protected static <ReqT, RespT> OutStreamObserver<ReqT> asyncBidiStreamingCall(
             ClientCall<ReqT, RespT> call,
             StreamObserver<RespT> responseObserver) {
         AsyncBidiStreamingOutAdapter<ReqT, RespT> adapter