+++ table/src/main/java/tech/ydb/table/utils/OperationParamUtils.java
@@ -2,6 +2,7 @@ package tech.ydb.table.utils;
 
 import com.google.protobuf.Duration;
 import tech.ydb.OperationProtos;
+import tech.ydb.common.CommonProtos.FeatureFlag.Status;
 import tech.ydb.table.settings.RequestSettings;
 
 public class OperationParamUtils {
@@ -13,6 +14,9 @@ public class OperationParamUtils {
         OperationProtos.OperationParams.Builder builder = OperationProtos.OperationParams.newBuilder();
         requestSettings.getOperationTimeout().ifPresent(duration -> builder.setOperationTimeout(toProto(duration)));
         requestSettings.getCancelAfter().ifPresent(duration -> builder.setCancelAfter(toProto(duration)));
+        requestSettings.getReportCostInfo().ifPresent(report ->
+                builder.setReportCostInfo(report ? Status.ENABLED : Status.DISABLED)
+        );
         return builder.build();
     }