+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.13.7 ##
+
+* Added regular re-discovery for new YDB transport implementation
+* Added node pessimization on receiving transport error in new YDB transport implementation
+
+## 1.13.6 ##
+
+* Added reportCostInfo flag to RequestSettings
+* Added costInfo to DataQueryResult
+
+## 1.13.5 ##
+
+* Use cached version of IAM Auth Provider
+
 ## 1.13.4 ##
 
 * Fix some issues in session pool